Remove uspport for `.packages` files.
Simplify (braking) API which no longer needs to support two files.

Since this is breaking anyway, add `final` to classes,
and make comparison operators on `LanguageVersion` instance methods.

Add tool to query package configuration for a file or directory:
`bin/package_config_of.dart`.
