+++ pkgs/package_config/README.md
@@ -21,6 +21,5 @@ The primary libraries of this package are
     Just the `PackageConfig` class and other types needed to use
     package configurations. This library does not depend on `dart:io`.
 
-The package includes deprecated backwards compatible functionality to
-work with the `.packages` file. This functionality will not be maintained,
-and will be removed in a future version of this package.
+The package no longer contains backwards compatible functionality to
+work with `.packages` files.
