+++ pkgs/package_config/lib/package_config_types.dart
@@ -17,6 +17,5 @@ export 'src/package_config.dart'
     show
         InvalidLanguageVersion,
         LanguageVersion,
-        LanguageVersionRelationalOperators,
         Package,
         PackageConfig;
