+++ pkgs/package_config/lib/src/package_config_io.dart
@@ -9,10 +9,8 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'errors.dart';
-import 'package_config_impl.dart';
+import 'package_config.dart';
 import 'package_config_json.dart';
-import 'packages_file.dart' as packages_file;
-import 'util.dart';
 import 'util_io.dart';
 
 /// Name of directory where Dart tools store their configuration.
@@ -25,32 +23,11 @@ const dartToolDirName = '.dart_tool';
 /// File is stored in the dart tool directory.
 const packageConfigFileName = 'package_config.json';
 
-/// Name of file containing legacy package configuration data.
-///
-/// File is stored in the package root directory.
-const packagesFileName = '.packages';
-
 /// Reads a package configuration file.
 ///
-/// Detects whether the [file] is a version one `.packages` file or
-/// a version two `package_config.json` file.
-///
-/// If the [file] is a `.packages` file and [preferNewest] is true,
-/// first checks whether there is an adjacent `.dart_tool/package_config.json`
-/// file, and if so, reads that instead.
-/// If [preferNewest] is false, the specified file is loaded even if it is
-/// a `.packages` file and there is an available `package_config.json` file.
-///
 /// The file must exist and be a normal file.
-Future<PackageConfig> readAnyConfigFile(
-    File file, bool preferNewest, void Function(Object error) onError) async {
-  if (preferNewest && fileName(file.path) == packagesFileName) {
-    var alternateFile = File(
-        pathJoin(dirName(file.path), dartToolDirName, packageConfigFileName));
-    if (alternateFile.existsSync()) {
-      return await readPackageConfigJsonFile(alternateFile, onError);
-    }
-  }
+Future<PackageConfig> readConfigFile(
+    File file, void Function(Object error) onError) async {
   Uint8List bytes;
   try {
     bytes = await file.readAsBytes();
@@ -58,38 +35,25 @@ Future<PackageConfig> readAnyConfigFile(
     onError(e);
     return const SimplePackageConfig.empty();
   }
-  return parseAnyConfigFile(bytes, file.uri, onError);
+  return parsePackageConfigBytes(bytes, file.uri, onError);
 }
 
-/// Like [readAnyConfigFile] but uses a URI and an optional loader.
-Future<PackageConfig> readAnyConfigFileUri(
+/// Like [readConfigFile] but uses a URI and an optional loader.
+Future<PackageConfig> readConfigFileUri(
     Uri file,
     Future<Uint8List?> Function(Uri uri)? loader,
-    void Function(Object error) onError,
-    bool preferNewest) async {
+    void Function(Object error) onError) async {
   if (file.isScheme('package')) {
     throw PackageConfigArgumentError(
         file, 'file', 'Must not be a package: URI');
   }
   if (loader == null) {
     if (file.isScheme('file')) {
-      return await readAnyConfigFile(File.fromUri(file), preferNewest, onError);
+      return await readConfigFile(File.fromUri(file), onError);
     }
     loader = defaultLoader;
   }
-  if (preferNewest && file.pathSegments.last == packagesFileName) {
-    var alternateFile = file.resolve('$dartToolDirName/$packageConfigFileName');
-    Uint8List? bytes;
-    try {
-      bytes = await loader(alternateFile);
-    } catch (e) {
-      onError(e);
-      return const SimplePackageConfig.empty();
-    }
-    if (bytes != null) {
-      return parsePackageConfigBytes(bytes, alternateFile, onError);
-    }
-  }
+
   Uint8List? bytes;
   try {
     bytes = await loader(file);
@@ -101,20 +65,6 @@ Future<PackageConfig> readAnyConfigFileUri(
     onError(PackageConfigArgumentError(
         file.toString(), 'file', 'File cannot be read'));
     return const SimplePackageConfig.empty();
-  }
-  return parseAnyConfigFile(bytes, file, onError);
-}
-
-/// Parses a `.packages` or `package_config.json` file's contents.
-///
-/// Assumes it's a JSON file if the first non-whitespace character
-/// is `{`, otherwise assumes it's a `.packages` file.
-PackageConfig parseAnyConfigFile(
-    Uint8List bytes, Uri file, void Function(Object error) onError) {
-  var firstChar = firstNonWhitespaceChar(bytes);
-  if (firstChar != $lbrace) {
-    // Definitely not a JSON object, probably a .packages.
-    return packages_file.parse(bytes, file, onError);
   }
   return parsePackageConfigBytes(bytes, file, onError);
 }
@@ -131,18 +81,6 @@ Future<PackageConfig> readPackageConfigJsonFile(
   return parsePackageConfigBytes(bytes, file.uri, onError);
 }
 
-Future<PackageConfig> readDotPackagesFile(
-    File file, void Function(Object error) onError) async {
-  Uint8List bytes;
-  try {
-    bytes = await file.readAsBytes();
-  } catch (error) {
-    onError(error);
-    return const SimplePackageConfig.empty();
-  }
-  return packages_file.parse(bytes, file.uri, onError);
-}
-
 Future<void> writePackageConfigJsonFile(
     PackageConfig config, Directory targetDirectory) async {
   // Write .dart_tool/package_config.json first.
@@ -153,14 +91,5 @@ Future<void> writePackageConfigJsonFile(
 
   var sink = file.openWrite(encoding: utf8);
   writePackageConfigJsonUtf8(config, baseUri, sink);
-  var doneJson = sink.close();
-
-  // Write .packages too.
-  file = File(pathJoin(targetDirectory.path, packagesFileName));
-  baseUri = file.uri;
-  sink = file.openWrite(encoding: utf8);
-  writeDotPackages(config, baseUri, sink);
-  var donePackages = sink.close();
-
-  await Future.wait([doneJson, donePackages]);
+  await sink.close();
 }
