+++ pkgs/package_config/lib/src/package_config_json.dart
@@ -8,8 +8,7 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'errors.dart';
-import 'package_config_impl.dart';
-import 'packages_file.dart' as packages_file;
+import 'package_config.dart';
 import 'util.dart';
 
 const String _configVersionKey = 'configVersion';
@@ -25,10 +24,6 @@ const List<String> _packageNames = [
   _packageUriKey,
   _languageVersionKey
 ];
-
-const String _generatedKey = 'generated';
-const String _generatorKey = 'generator';
-const String _generatorVersionKey = 'generatorVersion';
 
 final _jsonUtf8Decoder = json.fuse(utf8).decoder;
 
@@ -263,23 +258,6 @@ Map<String, Object?> packageConfigToJson(PackageConfig config, Uri? baseUri) =>
           }
       ],
     };
-
-void writeDotPackages(PackageConfig config, Uri baseUri, StringSink output) {
-  var extraData = config.extraData;
-  // Write .packages too.
-  String? comment;
-  if (extraData is Map<String, Object?>) {
-    var generator = extraData[_generatorKey];
-    if (generator is String) {
-      var generated = extraData[_generatedKey];
-      var generatorVersion = extraData[_generatorVersionKey];
-      comment = 'Generated by $generator'
-          "${generatorVersion is String ? " $generatorVersion" : ""}"
-          "${generated is String ? " on $generated" : ""}.";
-    }
-  }
-  packages_file.write(output, config, baseUri: baseUri, comment: comment);
-}
 
 /// If "extraData" is a JSON map, then return it, otherwise return null.
 ///
