+++ pkgs/package_config/pubspec.yaml
@@ -1,5 +1,5 @@
 name: package_config
-version: 2.2.0-wip
+version: 3.0.0-wip
 description: Support for reading and writing Dart Package Configuration files.
 repository: https://github.com/dart-lang/tools/tree/main/pkgs/package_config
 issue_tracker: https://github.com/dart-lang/tools/labels/package%3Apackage_config
@@ -11,5 +11,5 @@ dependencies:
   path: ^1.8.0
 
 dev_dependencies:
-  dart_flutter_team_lints: ^3.0.0
+  dart_flutter_team_lints: ^3.1.0
   test: ^1.16.0
