+++ pkgs/package_config/test/parse_test.dart
@@ -8,75 +8,11 @@ import 'dart:typed_data';
 import 'package:package_config/package_config_types.dart';
 import 'package:package_config/src/errors.dart';
 import 'package:package_config/src/package_config_json.dart';
-import 'package:package_config/src/packages_file.dart' as packages;
 import 'package:test/test.dart';
 
 import 'src/util.dart';
 
 void main() {
-  group('.packages', () {
-    test('valid', () {
-      var packagesFile = '# Generated by pub yadda yadda\n'
-          'foo:file:///foo/lib/\n'
-          'bar:/bar/lib/\n'
-          'baz:lib/\n';
-      var result = packages.parse(utf8.encode(packagesFile),
-          Uri.parse('file:///tmp/file.dart'), throwError);
-      expect(result.version, 1);
-      expect({for (var p in result.packages) p.name}, {'foo', 'bar', 'baz'});
-      expect(result.resolve(pkg('foo', 'foo.dart')),
-          Uri.parse('file:///foo/lib/foo.dart'));
-      expect(result.resolve(pkg('bar', 'bar.dart')),
-          Uri.parse('file:///bar/lib/bar.dart'));
-      expect(result.resolve(pkg('baz', 'baz.dart')),
-          Uri.parse('file:///tmp/lib/baz.dart'));
-
-      var foo = result['foo']!;
-      expect(foo, isNotNull);
-      expect(foo.root, Uri.parse('file:///foo/'));
-      expect(foo.packageUriRoot, Uri.parse('file:///foo/lib/'));
-      expect(foo.languageVersion, LanguageVersion(2, 7));
-      expect(foo.relativeRoot, false);
-    });
-
-    test('valid empty', () {
-      var packagesFile = '# Generated by pub yadda yadda\n';
-      var result = packages.parse(
-          utf8.encode(packagesFile), Uri.file('/tmp/file.dart'), throwError);
-      expect(result.version, 1);
-      expect({for (var p in result.packages) p.name}, <String>{});
-    });
-
-    group('invalid', () {
-      var baseFile = Uri.file('/tmp/file.dart');
-      void testThrows(String name, String content) {
-        test(name, () {
-          expect(
-              () => packages.parse(utf8.encode(content), baseFile, throwError),
-              throwsA(isA<FormatException>()));
-        });
-        test('$name, handle error', () {
-          var hadError = false;
-          packages.parse(utf8.encode(content), baseFile, (error) {
-            hadError = true;
-            expect(error, isA<FormatException>());
-          });
-          expect(hadError, true);
-        });
-      }
-
-      testThrows('repeated package name', 'foo:lib/\nfoo:lib\n');
-      testThrows('no colon', 'foo\n');
-      testThrows('empty package name', ':lib/\n');
-      testThrows('dot only package name', '.:lib/\n');
-      testThrows('dot only package name', '..:lib/\n');
-      testThrows('invalid package name character', 'f\\o:lib/\n');
-      testThrows('package URI', 'foo:package:bar/lib/');
-      testThrows('location with query', 'f\\o:lib/?\n');
-      testThrows('location with fragment', 'f\\o:lib/#\n');
-    });
-  });
-
   group('package_config.json', () {
     test('valid', () {
       var packageConfigFile = '''
